persister 관련 로직은 별도 함수로 이동
